+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Implement query combination: `union`, `intersect`, `except`
 * Implement `#not_in` method
 * Implement case expression (using `#cond`)
 * Implement not in expression